Move all icons to be under the same folder
