+++ plugins/backstage-plugin-env0/src/components/env0-environment-details-card/env0-environment-details-card.tsx
@@ -11,13 +11,13 @@ import {
 } from '@backstage/core-components';
 import isEmpty from 'lodash/isEmpty';
 import { getGitProvider, getShortenRepo } from './get-shorten-repo';
-import { VcsIcon } from './vcs-icon';
 import { CardHeader, styled, Button } from '@material-ui/core';
 import Cached from '@material-ui/icons/Cached';
 import { Env0Api, env0ApiRef } from '../../api';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { ENV0_ENVIRONMENT_ANNOTATION } from '../common/is-plugin-available';
-import { Env0Icon } from '../env0-icon';
+import { Env0Icon } from '../icons';
+import { VcsIcon } from './vcs-icon';
 
 type CardProps = {
   children: React.ReactNode;
