--- plugins/backstage-plugin-env0/src/components/env0-environment-details-card/vcs-icon/index.ts
@@ -1,1 +1,0 @@
-export {VcsIcon} from "./vcs-icon";
