--- plugins/backstage-plugin-env0/src/components/env0-icon/index.ts
@@ -1,1 +1,0 @@
-export {Env0Icon} from "./env0-icon";
