+++ plugins/backstage-plugin-env0/src/components/icons/env0-icon.tsx
