+++ plugins/backstage-plugin-env0/src/components/icons/index.tsx
@@ -1,2 +1,4 @@
 export * from './destroy-icon';
 export * from './deploy-icon';
+export * from './env0-icon';
+export * from './vcs-icon';
