+++ plugins/backstage-plugin-env0/src/components/icons/vcs-icon/bitbucket.tsx
