+++ plugins/backstage-plugin-env0/src/components/icons/vcs-icon/gitlab.tsx
