+++ plugins/backstage-plugin-env0/src/components/icons/vcs-icon/index.ts
@@ -1,0 +1,6 @@
+export * from './azure-devops';
+export * from './bitbucket';
+export * from './github';
+export * from './gitlab';
+export * from './helm';
+export * from './git';
